requirements.txt
ultralytics 
supervision 
pillow 
psutil
pystray